Get ASL sign image (ASL Signbank)

This function inputs an ASL ID gloss or number and downloads the image of the corresponding sign entry in ASL Signbank.

get_image_asl(
  id,
  glosstext = FALSE,
  gravity = "north",
  location = "+10+20",
  fontsize = 70
)

Arguments

id
ID number for the ASL sign

glosstext
Prints the ID gloss onto the image if set to `TRUE` (default = `FALSE`)

gravity
The anchor point of the text annotation if selected (default = "north")

location
The exact location of the text annotation relative to `gravity`

fontsize
The fontsize for the text annotation (default = 70)

Value

The path of the image file that was downloaded

Details

A few esthetics arguments for the text annotation are optional (e.g. `gravity`, `fontsize`)