Make ELAN image

This function creates image files (with or without text) from an ELAN annotation file

make_elan_image(
  path,
  destination = "./",
  segmentation_tier = "",
  gloss_tier = "",
  video = "",
  combine = TRUE,
  scale = 1,
  crop = FALSE,
  region = "100x150+100+100",
  text = "",
  textcolor = "white",
  font = "Helvetica",
  fontsize = "40",
  gravity = "southwest",
  location = "+20+20",
  border = FALSE
)

Arguments

path

The path to a single ELAN .eaf file or a directory with .eaf files

destination

The path/filename of the output image

segmentation_tier

List of specific tiers with segmentation annotations to be included

gloss_tier

List of specific tiers with gloss annotations to be included

video

Specify a video path to be used (if not, first linked video per eaf is used)

combine

Represent each sign by an overlay of the first and last frame of the sign (default=TRUE)

scale

The resizing proportion of the original image (0 to 1; default is 1 = original size)

crop

Whether to crop image to region (`crop=FALSE` is default)

region

The region to be cropped (defaults to '100x150+100+100', (width x height +upper_x +upper_y)). Only for `crop=TRUE`

text

The text string to be added to the image

textcolor

The color of the text annotation (default is "white")

font

The font of the text annotation (default is "Helvetica")

fontsize

The fontsize of the text annotation (default is 70)

gravity

The anchor point of the added text (default is "southwest")

location

The location of the text relative to the anchor point

border

Set to `TRUE` if you want a black border frame for the output image

Value

The name of the image file(s) created