Make image example from an image file

This function inputs an image file and allows for resizing, cropping, and text annotations

make_image_ex(
  file,
  destfile = "",
  crop = FALSE,
  scale = 1,
  region = "100x150+100+100",
  text = "",
  textcolor = "white",
  font = "Helvetica",
  fontsize = "70",
  gravity = "southwest",
  location = "+20+20",
  border = FALSE
)

Arguments

file
The path to the image to be modified

destfile
The specified path to the modified file created

crop
Whether to use automatic face detection or not (`automatic=TRUE` is default)

scale
The resizing proportion of the original image (0 to 1; default is 1 = original size)

region
The region to be cropped (defaults to '100x150+100+100', (width x height +upper_x +upper_y)). Only for `crop=TRUE`

text
The text string to be added to the image

textcolor
The color of the text annotation (default is "white")

font
The font of the text annotation (default is "Helvetica")

fontsize
The fontsize of the text annotation (default is 70)

gravity
The anchor point of the added text (default is "southwest")

location
The location of the text relative to the anchor point

border
Set to `TRUE` if you want a black border frame for the output image

Value

The name of the image file that was created